@@ -5,8 +5,8 @@ apply plugin: "idea"
 apply plugin: 'groovy'
 apply plugin: 'java-library-distribution'
 
-sourceCompatibility = "1.6"
-targetCompatibility = "1.6"
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
 
 group = "canfield.josh"
 version = "1.1"
